requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-dotenv==1.0.0
dj-rest-auth==5.0.1
djangorestframework-simplejwt==5.3.0
django-filter==23.5

